Separating Columns in the Table Component for Improved Code Readability.
